refactor
move is_xx_op from rnn_utils to utils
