+++ tf2onnx/rewriter/unit_rnn_rewriter_base.py
@@ -13,10 +13,9 @@ from onnx import TensorProto
 from tf2onnx.graph_builder import GraphBuilder
 from tf2onnx.rewriter.loop_rewriter_base import LoopRewriterBase, Context
 from tf2onnx.rewriter.rnn_utils import REWRITER_RESULT, get_pattern, \
-    get_rnn_scope_name, parse_rnn_loop, is_select_op, is_tensor_array_write_op, \
-    seq_len_pattern
+    get_rnn_scope_name, parse_rnn_loop, seq_len_pattern
+from tf2onnx.utils import is_select_op, is_tensor_array_write_op
 from tf2onnx.graph_matcher import GraphMatcher
-
 
 
 logger = logging.getLogger(__name__)
